third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/debugger-compile-and-run-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests separate compilation and run.


Running: testSuccessfulCompileAndRun
Compiling script
Running script
Script result: 3

Running: testRunError
Compiling script
Running script
exceptionDetails:
   Uncaught ReferenceError: c is not defined
    at test.js:1:16
   line: 0, column: 15
   exceptionDetails stack trace:
       url: test.js
       function: 
       line: 0

Running: testCompileError
Compiling script
exceptionDetails:
   Uncaught SyntaxError: Unexpected token }
   line: 0, column: 0
   no stack trace attached to exceptionDetails
